+++ src/app/login/page.tsx
@@ -35,7 +35,7 @@ const LoginPage = () => {
 
     if (signin_res?.ok && !signin_res.error) {
       setError(false);
-      router.push("/noten");
+      router.push("/grades");
       // router.push("/");
     } else {
       setError(true);
